refactor(): Reorganize the folders and move files in them
It's clearer on how to view the files.
